Add files via upload
ADDITION: INSTEAD of printing the result for the addition in the addition function, have the addition function simply return the result in $V0. Then have the main function print the answer.

SUBTRACTION: INSTEAD of printing the result for the subtraction in the subtraction function, have the subtraction function simply return the result in $V0. Then have the main function print the answer.

EVERYTHING ELSE should remain the same as 3a. You MUST still print the menu, obtain the user's choice, perform the action and continue looping (printing the menu each time) until the user chooses to exit.

Submit the file as a plain text, MIPSYM assembly file.